+++ Code/Commando/dlgevamaptab.cpp
@@ -121,7 +121,7 @@ EvaMapTabClass::On_Init_Dialog (void)
 				//
 				int color = VRGB_TO_INT32 (objective->Type_To_Color ());
 				int item_index = map_ctrl->Add_Marker (L"", objective->Position, RectClass (32, 0, 64, 32), color);
-				map_ctrl->Set_Marker_Data (item_index, (uint32)objective);
+				map_ctrl->Set_Marker_Data (item_index, (uintptr_t)objective);
 			}
 		}
