+++ Code/Commando/dlgevaviewertab.cpp
@@ -107,7 +107,7 @@ EvaViewerTabClass::On_Init_Dialog (void)
 			//
 			int item_index = ListCtrl->Insert_Entry (index, object.Get_Name ());
 			if (item_index >= 0) {
-				ListCtrl->Set_Entry_Data (item_index, 0, (uint32)&object);
+				ListCtrl->Set_Entry_Data (item_index, 0, (uintptr_t)&object);
 
 				//
 				//	Colorize the entry
